feat: jotai - product 아톰 생성 & 엔티티별 의존성 분리
