+++ src/advanced/atoms/index.ts
@@ -1,2 +1,4 @@
 export * from "./notifications";
 export * from "./cart";
+export * from "./coupons";
+export * from "./products";
